+++ browser/activator.js
@@ -4,6 +4,7 @@ var raf = require('raf');
 var clone = require('./clone');
 var emitter = require('./emitter');
 var fetcher = require('./fetcher');
+var prefetcher = require('./prefetcher');
 var view = require('./view');
 var router = require('./router');
 var state = require('./state');
@@ -47,6 +48,7 @@ function go (url, options) {
   }
 
   global.DEBUG && global.DEBUG('[activator] fetching %s', route.url);
+  prefetcher.abortIntents();
   fetcher.abortPending();
   fetcher(route, { element: context, source: 'intent' }, maybeResolved);