+++ browser/prefetcher.js
@@ -1,0 +1,64 @@
+'use strict';
+
+var state = require('./state');
+var router = require('./router');
+var fetcher = require('./fetcher');
+var activator = require('./activator');
+var intents = [];
+var jobs = [];
+
+function busy (url) {
+  return jobs.indexOf(url) !== -1;
+}
+
+function registerIntent (url) {
+  if (intents.indexOf(url) === -1) {
+    intents.push(url);
+  }
+}
+
+function abortIntent (url) {
+  intents.splice(intents.indexOf(url), 1);
+}
+
+function abortIntents () {
+  intents = [];
+}
+
+function start (url, element) {
+  if (state.cache !== true) { // can't prefetch if caching is disabled
+    return;
+  }
+  if (intents.length) { // don't prefetch if the human wants to navigate: it'd abort the previous attempt
+    return;
+  }
+  var route = router(url);
+  if (route === null) { // only prefetch taunus view routes
+    return;
+  }
+  if (busy(url)) { // already prefetching this url
+    return;
+  }
+
+  global.DEBUG && global.DEBUG('[prefetcher] prefetching %s', route.url);
+  jobs.push(url);
+  fetcher(route, { element: element, source: 'prefetch' }, fetched);
+
+  function fetched () {
+    jobs.splice(jobs.indexOf(url), 1);
+    if (intents.indexOf(url) !== -1) {
+      abortIntent(url);
+
+      global.DEBUG && global.DEBUG('[prefetcher] resumed navigation for %s', route.url);
+      activator.go(route.url, { context: element });
+    }
+  }
+}
+
+module.exports = {
+  busy: busy,
+  start: start,
+  registerIntent: registerIntent,
+  abortIntent: abortIntent,
+  abortIntents: abortIntents
+};